Alway initialize the config section
The /lib/gluon/upgrade/500-ssid-changer script is called when the device is
initialized, reconfigured or upgraded. It will set the configuration to the
site content and prepare everything for the actual gluon_ssid_changer
script.

But it will check whether the ssid-changer.settings.enabled entry exists.
If it does, then it will just ignore the values from the site and skip the
initialization. But the problem here is that the package already provides
this entry in his standard initialization. The section will therefore never
get initialized and thus the package is disabled on all installations.

Instead just copy the content of ssid-changer.settings.enabled when it
exists and otherwise use the data from the site. And the rest of the
configuration must always be copied from the site.